Refactor agent runtime interactions and modularize handler functions
- Introduced new modules for handlers, helpers, and interactions to improve organization and maintainability of the agent runtime code.
- Updated interaction event handling to streamline session and turn management, ensuring better encapsulation of logic.
- Enhanced error handling and logging within interaction functions to improve traceability during session lifecycle events.
- Added comprehensive tests to validate the new modular structure and interaction recording processes, ensuring robust functionality across the agent runtime.
